+++ website/src/pages/index.js
@@ -64,7 +64,7 @@ function Home() {
                 'button button--outline button--secondary button--lg',
                 styles.getStarted,
               )}
-              to={useBaseUrl('docs/index')}
+              to={useBaseUrl('docs')}
             >
               Documentation
             </Link>