+++ packages/apollo-angular/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support Apollo Client 3.0 (use `@apollo/client/core`)
 - `Http`, `HttpBatch` and `Headers` links are now part of apollo-angular: `apollo-angular/http` and `apollo-angular/headers`
 - Re-exports `gql` of `graphql-tag` (`import { gql } from 'apollo-angular'`)
+- Add `flushData` method to `TestingOperation` ([PR #1474](https://github.com/kamilkisiela/apollo-angular/pull/1474) by [@fetis](https://github.com/fetis))
 
 ### v1.10.0